Change int to big int on costumer cedula

+++ DataAccess/DTO/CostumerDTO.cs
@@ -11,7 +11,7 @@ namespace DataAccess.DTO
     public class CostumerDTO
     {
         public int id { get; set; }
-        public int cedulaJuridica { get; set; }
+        public long cedulaJuridica { get; set; }
         public string name { get; set; }
         public string province { get; set; }
         public string canton { get; set; }
